fix: extract S3 handlers to s3_handlers.rs + into_any in settings view
CI surfaced two compile-time issues with the v0.3.32 PR:

1. crates/rs-api/src/handlers.rs hit 1003 lines after adding the new
   clear_event_s3_chunks and get_s3_usage handlers, exceeding the
   1000-line per-file cap. Extracted both handlers (and the
   ClearChunksResponse / S3UsageEntry / S3UsageResponse types) into a
   new s3_handlers.rs module. Router updated to point at it.

2. leptos-ui settings view had an if/else chain returning different
   view!{} types — Leptos generates a unique type per view! macro call
   so the arms cannot unify directly. Added .into_any() to each branch.
