+++ crates/rs-api/src/lib.rs
@@ -9,6 +9,7 @@ pub mod obs;
 pub mod router;
 #[cfg(test)]
 mod router_tests;
+pub mod s3_handlers;
 pub mod state;
 pub mod stream_handlers;
 pub mod template_handlers;
